Feature: Upgrade terrain generator, bomb drop and explosion done, chest done and door done.

+++ Assets/Resources/Scripts/Tokens/Enemys/LifeBarEnemyController.cs
@@ -6,13 +6,13 @@ public class LifeBarEnemyController : MonoBehaviour
     public RectTransform barLife;
     public RectTransform barFront;
 
-    public void SetUpLife(int life)
+    public void SetUpLife(float life)
     {
         barLife.sizeDelta = new Vector2(0.15f * life, 0.12f);
         gameObject.SetActive(false);
     }
     
-    public void ChangeLife(int life, int maxLife)
+    public void ChangeLife(float life, float maxLife)
     {
         barFront.sizeDelta = new Vector2(0.15f * (maxLife - life), 0.12f);
     }
